Ensure compatability with Django 4.1
By moving the AppConfig to the correct location and deprecating
usage of the `default_app_config` attribute to those versions of
Django less than 3.2 when it was officially deprecated.
